arch/x86/entry/entry_64.S: Removed the fast path in a syscall entry
Our syscall logger should record both of the fast path and the slow
path. I have implmented syscall logger in a slow path. Instead of
implementing same thing in the fast path, to reduce my effort, I
removed the fast path because he fast path will be removed later
because of PTI. This is clearly fine when we focus on only our
reproducer's capability, but it makes kernel slow. It would better not
to conduct any performance measurement in this version.