mm/logger: Handle correctly that a kernel dies before syscall returns
I was wrong. All data except ->exit_time should be written when a
syscall enter to handle the situation that a kernel dies before the
syscall returns. Revert to the previous implementation.

But still this is buggy. There are some syscalls that are alive for a
few seconds (i.e., epoll_wait), and during their lifetime, it is
likely that the ring buffer is filled entirely and start a next
round. In this case, bad_happened will be true. I will fix this by
keeping the entry and release it when the syscall returns.